requirements.txt
geemap>=0.21.0
leafmap
mapwidget
solara
geopandas
pydantic<2.0
huggingface_hub